@@ -10,16 +10,16 @@
     "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
   },
   "devDependencies": {
-    "axios": "^0.18",
-    "cross-env": "^5.1.4",
-    "laravel-mix": "^2.1",
-    "laravel-mix-merge-manifest": "^0.1.1",
-    "jquery": "^3.2",
-    "vue": "^2.1.10",
-    "flatpickr": "^4.4.6"
+    "axios": "^0.19.0",
+    "cross-env": "^6.0.3",
+    "flatpickr": "^4.6.3",
+    "jquery": "^3.4.1",
+    "laravel-mix": "^5.0.0",
+    "laravel-mix-merge-manifest": "^0.1.2",
+    "vue": "^2.6.10"
   },
   "dependencies": {
-    "@babel/polyfill": "^7.4.4",
+    "@babel/polyfill": "^7.7.0",
     "tooltip.js": "^1.3.1",
     "url-polyfill": "^1.1.5",
     "url-search-params-polyfill": "^6.0.0",
